Dockerfile + ejs
The file slowly start to take direction. I am able to deploy the project image on the docker .
In this way I can proof view counts for each client. this will tackle 2 problems. Different players on different clientss . And The clients can display the number of  games , replacing the number of visits.