DB Conection
I created a simple mongo DB  and  instert couple of users for testing. Now i can acces the log in from the DB instead of the List.
Trough a websocket connection. This recive the data on the JS authenticate the user credentials and send the data to the app.js  on the main HTML . retriving username and data of games