Registration SUccesful & docker update
Now registration is working I implement a new .ejs & .jsas  when i try to implement the registration on the app.js it was not letting me use it as it was taking inf oof other buttons. So i create a new regist.js to retrive data of thje button and send it to the index.js and then print data resultant either succes reg or no. Test deploy the uplodade project on docker containers and its running smothly connecting info correctly on the DB